% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.dgCMatrix.R
\name{assert_dgCMatrix}
\alias{assert_dgCMatrix}
\title{Check if the sparse matrix valid}
\usage{
assert_dgCMatrix(adj_matrix, non_negative = TRUE)
}
\arguments{
\item{adj_matrix}{The sparse matrix to check}

\item{non_negative}{Check if any non negative vlues}
}
\description{
Check if the sparse matrix valid
}
