% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusr-package.R
\docType{package}
\name{diffusr-package}
\alias{diffusr-package}
\title{Package: diffusr}
\description{
Network diffusion algorithms in R.
}
\author{
Simon Dirmeier | Computational Biology Group, ETH ZURICH | \email{simon.dirmeier@bsse.ethz.ch}
}
\keyword{package}

