% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_util.R
\name{hub.correction}
\alias{hub.correction}
\title{Correct for hubs in an adjacency matrix}
\usage{
hub.correction(obj)
}
\arguments{
\item{obj}{\code{\link[base]{matrix}} (or
\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}) for which hubs
are corrected}
}
\value{
returns the \code{\link[base]{matrix}} (or
  \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}) with hub correction
}
\description{
Correct for hubs in an adjacency matrix
}
\examples{
W <- matrix(abs(rnorm(10000)), 100, 100)
cor.hub <- hub.correction(W)
}
