% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_util.R
\name{laplacian}
\alias{laplacian}
\title{Calculate the Laplacian matrix for a matrix}
\usage{
laplacian(obj, ...)
}
\arguments{
\item{obj}{matrix for which the Laplacian is calculated}

\item{...}{additional params}
}
\value{
returns the Laplacian
}
\description{
Calculate the Laplacian matrix for a matrix
}
\examples{
W <- matrix(abs(rnorm(10000)), 100, 100)
lapl.W <- laplacian(W)
}
\author{
Simon Dirmeier, email{simon.dirmeier@bsse.ethz.ch}
}

