% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_util.R
\name{normalize.laplacian}
\alias{normalize.laplacian}
\title{Calculate the Laplacian of a matrix}
\usage{
normalize.laplacian(obj, ...)
}
\arguments{
\item{obj}{matrix for which the Laplacian is calculated}

\item{...}{additional params}
}
\value{
returns the Laplacian
}
\description{
Calculate the Laplacian of a matrix
}
\examples{
W <- matrix(abs(rnorm(10000)), 100, 100)
lapl.W <- normalize.laplacian(W)
}
\author{
Simon Dirmeier, \email{simon.dirmeier@gmx.de}
}

