% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_util.R
\name{normalize}
\alias{normalize}
\title{Calculate a stochastic column normalized matrix}
\usage{
normalize(obj, ...)
}
\arguments{
\item{obj}{matrix for which the normalized stochastic matrix is created}

\item{...}{additional params}
}
\value{
returns the normalized matrix
}
\description{
Calculate a stochastic column normalized matrix
}
\examples{
W <- matrix(abs(rnorm(10000)), 100, 100)
stoch.W <- normalize(W)
}
\author{
Simon Dirmeier, email{simon.dirmeier@bsse.ethz.ch}
}

