% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrw.R
\name{random.walk}
\alias{random.walk}
\title{Graph diffusion using a Markov random walk}
\usage{
random.walk(p0, graph, r = 0.5, ...)
}
\arguments{
\item{p0}{the starting distribution of the Markov chain (does not need to sum to one)}

\item{graph}{a non-negative matrix}

\item{r}{the restart probability if a Markov random walk with restart is desired}

\item{...}{additional params}
}
\value{
returns the stationary distribution as vector
}
\description{
A Markov Random Walk takes an inital distribution \code{p0} and calculates the stationary distribution of that.
The diffusion process is regulated by a restart probability \code{r}.
It basically controls how often the MRW jumps back to the initial values.
}
\examples{
# count of nodes
n <- 5
# starting distribution (has to sum to one)
p0    <- as.vector(rmultinom(1, 1, prob=rep(.2, n)))
# adjacency matrix (either normalized or not)
graph <- matrix(abs(rnorm(n*n)), n, n)
# computation of stationary distribution
pt    <- random.walk(p0, graph)
}
\author{
Simon Dirmeier, \email{simon.dirmeier@gmx.de}
}
\references{
Tong, H., Faloutsos, C., & Pan, J. Y. (2006),
Fast random walk with restart and its applications.\cr \cr
Koehler, S., Bauer, S., Horn, D., & Robinson, P. N. (2008),
Walking the interactome for prioritization of candidate disease genes.
\emph{The American Journal of Human Genetics}\cr \cr
}

