% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusr-package.R
\docType{package}
\name{diffusr-package}
\alias{diffusr-package}
\title{diffusr}
\description{
Network diffusion algorithms in R.
}
\author{
Simon Dirmeier \email{simon.dirmeier@gmx.de}
}
\references{
Tong, H., Faloutsos, C., & Pan, J. Y. (2006),
Fast random walk with restart and its applications.\cr \cr
Koehler, S., Bauer, S., Horn, D., & Robinson, P. N. (2008),
Walking the interactome for prioritization of candidate disease genes.
\emph{The American Journal of Human Genetics}\cr \cr
Bonacich, P. (1987),
Power and centrality: A family of measures.
\emph{American journal of sociology}\cr \cr
Leiserson, M. D., Vandin, F., Wu, H. T., Dobson, J. R., Eldridge, J. V., Thomas, J. L., ... & Lawrence, M. S. (2015),
Pan-cancer network analysis identifies combinations of rare somatic mutations across pathways and protein complexes.
\emph{Nature genetics}\cr \cr
\url{https://en.wikipedia.org/wiki/Laplacian_matrix} \cr
\url{https://en.wikipedia.org/wiki/Heat_equation}
}
\keyword{package}

