% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusr-package.R
\docType{package}
\name{diffusr-package}
\alias{diffusr-package}
\title{diffusr}
\description{
Network diffusion algorithms in R.
}
\author{
Simon Dirmeier \email{simon.dirmeier@gmx.de}
}
\keyword{package}

