% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_kernel.R
\name{heat.diffusion}
\alias{heat.diffusion}
\title{Do graph diffusion using heat diffusion}
\usage{
heat.diffusion(v, graph, t = 100, ...)
}
\arguments{
\item{v}{the starting distribution of the Markov chain}

\item{graph}{a weighted \code{igraph} object}

\item{t}{the restart probability if a Markov random walk with restart is desired}

\item{...}{additional params}
}
\value{
returns the heat on every node of the graph
}
\description{
Do graph diffusion using heat diffusion
}
\author{
Simon Dirmeier, email{simon.dirmeier@bsse.ethz.ch}
}

