% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_kernel.R
\name{heat.diffusion}
\alias{heat.diffusion}
\title{Do graph diffusion using an insulated heat kernel.}
\usage{
heat.diffusion(h, graph, r = 0.5, ...)
}
\arguments{
\item{h}{the starting heat distribution}

\item{graph}{a non-negative matrix}

\item{r}{'rate' of heat diffusion, where 1 is the maximum diffusion and 0 no diffusion at al}

\item{...}{additional params}
}
\value{
returns the heat on every node as a vector
}
\description{
Do graph diffusion using an insulated heat kernel.
}
\examples{
# count of nodes
n <- 5
# starting distribution (has to sum to one)
h    <- rmultinom(1, 1, prob=rep(.2, n))
# adjacency matrix (either normalized or not)
graph <- matrix(abs(rnorm(n*n)), n, n)
# computation of stationary distribution
pt    <- heat.diffusion(h, graph)
}
\author{
Simon Dirmeier, \email{simon.dirmeier@gmx.de}
}

