% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_kernel.R
\name{heat.diffusion}
\alias{heat.diffusion}
\title{Do graph diffusion using an insulated heat kernel.}
\usage{
heat.diffusion(h, graph, r = 0.5, ...)
}
\arguments{
\item{h}{the starting heat distribution}

\item{graph}{a non-negative matrix}

\item{r}{the restart probability of the heat kernel}

\item{...}{additional params}
}
\value{
returns the heat on every node as an
}
\description{
Do graph diffusion using an insulated heat kernel.
}
\author{
Simon Dirmeier, email{simon.dirmeier@bsse.ethz.ch}
}

