% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{Finds the k-NN nearest neighbors of every node in a graph}
\usage{
knn(nodes, graph, k = 1, ...)
}
\arguments{
\item{nodes}{vector of nodes for which the k-NN algorithm is applied  is applied}

\item{graph}{an <code>igraph</code> object}

\item{k}{the depth of the nearest neighbor search}
}
\value{
the kNN graph as an an <code>igraph</code> object
}
\description{
Finds the k-NN nearest neighbors of every node in a graph
}
\examples{
\dontrun{
 TODO
}
}
\author{
Simon Dirmeier
}

