% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{neighbors}
\alias{neighbors}
\title{Find the closest neighbors of a group of nodes in a graph.}
\usage{
neighbors(nodes, graph, k = 1L, use.edge.weights = F, ...)
}
\arguments{
\item{nodes}{vector of node indexes (1-based) for which the algorithm is applied iteratively}

\item{graph}{a non-negative matrix}

\item{k}{the depth of the nearest neighbor search}

\item{use.edge.weights}{boolean flags if the edge weights should be considered when doing nearest neighbor lookup}

\item{...}{additional params}
}
\value{
returns the kNN nodes as integer vector of node indexes
}
\description{
Find the closest neighbors of a group of nodes in a graph.
}
\examples{
 # count of nodes
 n <- 10
 # indexes (integer) of nodes for which neighbors should be searched
 node.idxs <- c(1L, 5L)
 # the adjaceny matrix (does not need to be symmetric)
 graph <- rbind(cbind(0, diag(n-1)), 0)
 # compte the neighbors until depth 3
 neighs <- neighbors(node.idxs, graph, 3)
}
\author{
Simon Dirmeier, email{simon.dirmeier@bsse.ethz.ch}
}

