% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{neighbors}
\alias{neighbors}
\title{Find the closest neighbors of a group of nodes in a graph.}
\usage{
neighbors(node.idxs, graph, k = 1L, use.edge.weights = F, ...)
}
\arguments{
\item{node.idxs}{vector of node indexes (1-based) for which the algorithm is applied iteratively}

\item{graph}{an <code>igraph</code> object}

\item{k}{the depth of the nearest neighbor search}

\item{use.edge.weights}{boolean flags if the edge weights should be considered when doing nearest neighbor lookup}

\item{...}{additional params}
}
\value{
returns the kNN graph as an an <code>igraph</code> object
}
\description{
Find the closest neighbors of a group of nodes in a graph.
}
\examples{
\dontrun{
 TODO
}
}
\author{
Simon Dirmeier
}

