% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{neighbors}
\alias{neighbors}
\title{Find the closest neighbors of a set nodes in a graph.}
\usage{
neighbors(nodes, graph, k = 1L, ...)
}
\arguments{
\item{nodes}{vector of node indexes (1-based) for which the algorithm is applied iteratively}

\item{graph}{a non-negative matrix}

\item{k}{the depth of the nearest neighbor search}

\item{...}{additional params}
}
\value{
returns the kNN nodes as list of integer vectors of node indexes
}
\description{
Runs Disjktra's algorithm to find the closest neighbors (shortest paths) for a set of nodes.
For every node in the set Disjktra's algorithm is run against all other nodes in the graph. Then the \code{k} nearest neighbors or chosen.
For settings where there are more than \code{k} neighbors having the same distance, all neighbors are returned.
}
\examples{
 # count of nodes
 n <- 10
 # indexes (integer) of nodes for which neighbors should be searched
 node.idxs <- c(1L, 5L)
 # the adjaceny matrix (does not need to be symmetric)
 graph <- rbind(cbind(0, diag(n-1)), 0)
 # compte the neighbors until depth 3
 neighs <- neighbors(node.idxs, graph, 3)
}
\author{
Simon Dirmeier, \email{simon.dirmeier@gmx.de}
}

