% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_util.R
\name{normalize}
\alias{normalize}
\title{Create a stochastically normalized matrix/vector}
\usage{
normalize(obj, ...)
}
\arguments{
\item{obj}{matrix/vector that is stochstically normalized}

\item{...}{additional params}
}
\value{
returns the normalized matrix/vector
}
\description{
Create a stochastically normalized matrix/vector
}
\examples{
W <- matrix(abs(rnorm(10000)), 100, 100)
stoch.W <- normalize(W)
}
\author{
Simon Dirmeier, \email{simon.dirmeier@gmx.de}
}

