% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrw.R
\name{random.walk}
\alias{random.walk}
\title{Finds the k-NN nearest neighbors of every node in a graph}
\usage{
random.walk(nodes, graph, r, ...)
}
\arguments{
\item{nodes}{vector of nodes for which the k-NN algorithm is applied  is applied}

\item{graph}{an <code>igraph</code> object}

\item{r}{the restart probability if a Markov random walk with restart is desired}
}
\value{
the kNN graph as an an <code>igraph</code> object
}
\description{
Finds the k-NN nearest neighbors of every node in a graph
}
\examples{
\dontrun{
 TODO
}
}
\author{
Simon Dirmeier
}

