% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrw.R
\name{random.walk}
\alias{random.walk}
\title{Do a Markov random walk on a graph.}
\usage{
random.walk(p0, graph, r = 0.5, ...)
}
\arguments{
\item{p0}{the starting distribution of the Markov chain}

\item{graph}{a non-negative matrix}

\item{r}{the restart probability if a Markov random walk with restart is desired}

\item{...}{additional params}
}
\value{
returns the stationary distribution as vector
}
\description{
Do a Markov random walk on a graph.
}
\examples{
# count of nodes
n <- 5
# starting distribution (has to sum to one)
p0    <- rmultinom(1, 1, prob=rep(.2, n))
# adjacency matrix (either normalized or not)
graph <- matrix(abs(rnorm(n*n)), n, n)
# computation of stationary distribution
pt    <- random.walk(p0, graph)
}
\author{
Simon Dirmeier, \email{simon.dirmeier@gmx.de}
}

