% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrw.R
\name{random.walk}
\alias{random.walk}
\title{Do a Markov random walk on a graph}
\usage{
random.walk(p0, graph, r, ...)
}
\arguments{
\item{p0}{the staring distribution for the Markov chain}

\item{graph}{a weighted <code>igraph</code> object}

\item{r}{the restart probability if a Markov random walk with restart is desired}

\item{...}{additional params}
}
\value{
returns the stationary distribution of the
}
\description{
Do a Markov random walk on a graph
}
\examples{
\dontrun{
 TODO
}
}
\author{
Simon Dirmeier
}

